+++ backend/server.js
@@ -10,6 +10,7 @@ import publishingRouter from './routes/publishing.js';
 import scheduledPostsRouter from './routes/scheduledPosts.js';
 import settingsRouter from './routes/settings.js';
 import redditRouter from './routes/reddit.js';
+import trackedPostsRouter from './routes/trackedPosts.js';
 import { urlHandler, callbackHandler } from './routes/instagramAuth.js';
 import { processScheduledPosts } from './utils/scheduler.js';
 import { autoRefreshInstagramTokens } from './services/instagram.js';
@@ -54,6 +55,7 @@ api.route('/', publishingRouter);
 api.route('/', scheduledPostsRouter);
 api.route('/', settingsRouter);
 api.route('/', redditRouter);
+api.route('/', trackedPostsRouter);
 
 // Instagram OAuth URL generator - protected (needs userId to create state token)
 api.get('/auth/instagram/url', urlHandler);